ble: Adding initial version of BLE Demo Apps in binary form for Android
Demo apps provided as .apk files which can be installed on Android phones.
Tested with Android Lollipop v5.1

To be used with example sketches provided in CurieBle library for AtlasEdge

Note that the AutomationIO apps are provided in 2 versions, one for ARM-based
phones and one for X86-based phones.  Please choose the appropriate .apk for
your device.

Source code for these binaries will follow in a later commit.

Signed-off-by: Dan O'Donovan <danielx.odonovan@intel.com>